Move `on` documentation to source but keep it displaying under @ember/helper keywords
